Revert "Search screen (#3)"
This reverts commit 3719c205e904b28bd592a13791a0dc59f361154b.
