+++ Popular Movies/Movie Cell/MovieCellProps.swift
